UART: Correcting private file

+++ lib/UART_private.h
@@ -1,9 +1,9 @@
 #ifndef _UART_PRIVATE_H
 #define _UART_PRIVATE_H
 
-#define UCSRB * ((u32*) 0x2A)
-#define UCSRA * ((u32*) 0x2B)
-#define UBRRL * ((u32*) 0x29)
-#define UDR * ((u32*)) 0x2C)
+#define UCSRB * ((u8*) 0x2A)
+#define UCSRA * ((u8*) 0x2B)
+#define UBRRL * ((u8*) 0x29)
+#define UDR * ((u8*)) 0x2C)
 
 #endif
